+++ src/init.c
@@ -13,6 +13,12 @@ extern SEXP serialize_to_file_(SEXP x_, SEXP filename_, SEXP serialize_opts_);
 extern SEXP parse_ndjson_file_as_df_  (SEXP filename_, SEXP nread_, SEXP nskip_, SEXP nprobe_, SEXP parse_opts_);
 extern SEXP parse_ndjson_file_as_list_(SEXP filename_, SEXP nread_, SEXP nskip_,               SEXP parse_opts_);
 
+extern SEXP serialize_df_to_ndjson_str_ (SEXP robj_,                 SEXP serialize_opts_);
+extern SEXP serialize_df_to_ndjson_file_(SEXP robj_, SEXP filename_, SEXP serialize_opts_);
+
+extern SEXP serialize_list_to_ndjson_str_ (SEXP robj_,                 SEXP serialize_opts_);
+extern SEXP serialize_list_to_ndjson_file_(SEXP robj_, SEXP filename_, SEXP serialize_opts_);
+
 static const R_CallMethodDef CEntries[] = {
   {"serialize_to_str_" , (DL_FUNC) &serialize_to_str_ , 2},
   {"serialize_to_file_", (DL_FUNC) &serialize_to_file_, 3},
@@ -23,6 +29,12 @@ static const R_CallMethodDef CEntries[] = {
   
   {"parse_ndjson_file_as_df_"  , (DL_FUNC) &parse_ndjson_file_as_df_  , 5},
   {"parse_ndjson_file_as_list_", (DL_FUNC) &parse_ndjson_file_as_list_, 4},
+  
+  {"serialize_df_to_ndjson_str_" , (DL_FUNC) &serialize_df_to_ndjson_str_ , 2},
+  {"serialize_df_to_ndjson_file_", (DL_FUNC) &serialize_df_to_ndjson_file_, 3},
+  
+  {"serialize_list_to_ndjson_str_" , (DL_FUNC) &serialize_list_to_ndjson_str_ , 2},
+  {"serialize_list_to_ndjson_file_", (DL_FUNC) &serialize_list_to_ndjson_file_, 3},
   
   {NULL , NULL, 0}
 };