+++ NEWS.md
@@ -1,4 +1,8 @@
 
+# yyjsonr 0.1.2  2023-08-20
+
+* Added `to_ndjson_file()` and `to_ndjson_str()`
+
 # yyjsonr 0.1.1  2023-08-19
 
 * Update to option setting to remove overhead